+++ src/test/java/roomescape/web/controller/ReservationControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.integration;
+package roomescape.web.controller;
 
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.startsWith;
@@ -6,14 +6,10 @@ import static org.hamcrest.Matchers.startsWith;
 import java.time.LocalDate;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.context.jdbc.Sql;
 import roomescape.service.request.ReservationRequest;
+import roomescape.support.RestAssuredTestSupport;
 
 /*
  * 테스트 데이터베이스 초기 데이터
@@ -21,17 +17,7 @@ import roomescape.service.request.ReservationRequest;
  * {ID=2, NAME=엘라, DATE=2024-05-04, TIME={ID=2, START_AT="11:00"}}
  * {ID=3, NAME=릴리, DATE=2023-08-05, TIME={ID=2, START_AT="11:00"}}
  */
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-@Sql(scripts = "/reset_test_data.sql")
-class ReservationIntegrationTest {
-
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-    }
+class ReservationControllerTest extends RestAssuredTestSupport {
 
     @Test
     @DisplayName("전체 예약 목록을 조회한다.")
