+++ src/test/java/roomescape/web/controller/ReservationTimeControllerTest.java
@@ -1,35 +1,21 @@
-package roomescape.integration;
+package roomescape.web.controller;
 
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.startsWith;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.context.jdbc.Sql;
 import roomescape.service.request.ReservationTimeRequest;
+import roomescape.support.RestAssuredTestSupport;
 
 /*
  * 테스트 데이터베이스 초기 데이터
  * {ID=1, START_AT=10:00}
  * {ID=2, START_AT=11:00}
  */
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-@Sql(scripts = "/reset_test_data.sql")
-class ReservationTimeIntegrationTest {
-
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-    }
+class ReservationTimeControllerTest extends RestAssuredTestSupport {
 
     @Test
     @DisplayName("전체 시간 목록을 조회한다.")
